// src/api/generated/models/filmSchedule.ts
/**
 * Generated by orval v8.1.0 🍺
 * Do not edit manually.
 * shift backend 🔥
 * Апи для выполнения индивидуальных заданий
 * OpenAPI spec version: 1.0
 */
import type { FilmScheduleSeance } from './filmScheduleSeance';

export type FilmSchedule = {
  /** Дата */
  date: string;
  /** Сеансы */
  seances: FilmScheduleSeance[];
};