// src/api/generated/models/filmScheduleSeance.ts
/**
 * Generated by orval v8.1.0 🍺
 * Do not edit manually.
 * shift backend 🔥
 * Апи для выполнения индивидуальных заданий
 * OpenAPI spec version: 1.0
 */
import type { FilmHall } from './filmHall';

export type FilmScheduleSeance = {
  /** Время сеанса */
  time: string;
  /** Зал */
  hall: FilmHall;
};